minor cli cleanups
