+++ scp/cli.py
@@ -1,98 +1,39 @@
-import sys
-import argparse
+import click
+from sys import stderr
+from scp import water, ethyl_alcohol, ethylene_glycol, propylene_glycol, methyl_alcohol
 
-from scp.ethyl_alcohol import EthylAlcohol
-from scp.ethylene_glycol import EthyleneGlycol
-from scp.methyl_alcohol import MethylAlcohol
-from scp.propylene_glycol import PropyleneGlycol
-from scp.water import Water
+# Add this to .bashrc for nice completion:  eval "$(_SCPROP_COMPLETE=bash_source scprop)"
 
+fluids = {
+    'water': water.Water,
+    'ethyl_alcohol': ethyl_alcohol.EthylAlcohol,
+    'ethylene_glycol': ethylene_glycol.EthyleneGlycol,
+    'methyl_alcohol': methyl_alcohol.MethylAlcohol,
+    'propylene_glycol': propylene_glycol.PropyleneGlycol
+}
+fluid_options = click.Choice(list(fluids.keys()))
+properties = ('viscosity', 'specific_heat', 'density', 'conductivity', 'prandtl', 'thermal_diffusivity')
+prop_options = click.Choice(properties)
+x_range = click.FloatRange(min=0.0, max=1.0)
 
-_valid_prop_strs = ["V", "C", "D", "L"]
-_valid_fluid_strs = ["EA", "EG", "MA", "PG", "W"]
-
-
-def _check_valid_prop(prop_str: str):
-    """
-    Worker function to validate property type strings
-    """
-    if prop_str not in _valid_prop_strs:
-        msg = f"PropStr not valid: {prop_str}"
-        raise ValueError(msg)
-
-
-def _check_valid_fluid(fluid_str: str):
-    """
-    Worker function to validate fluid type string
-    """
-    if fluid_str not in _valid_fluid_strs:
-        msg = f"FluidStr not valid: {fluid_str}"
-        raise ValueError(msg)
-
-
-def SCProps(prop_str, temp, x_conc, fluid_str):
-    """
-    Convenience function for console usage.
-    Developers should integrate the classes directly for typical applications.
-
-    @param prop_str: property type string
-    @param temp: temperature, in Celsius
-    @param x_conc: mixture concentration fraction, from 0 to 1
-    @param fluid_str: fluid type string
-    """
-
-    prop_str = prop_str.upper()
-    fluid_str = fluid_str.upper()
-
-    _check_valid_prop(prop_str)
-    _check_valid_fluid(fluid_str)
-
-    if fluid_str == "EA":
-        inst = EthylAlcohol(x_conc)
-    elif fluid_str == "EG":
-        inst = EthyleneGlycol(x_conc)
-    elif fluid_str == "MA":
-        inst = MethylAlcohol(x_conc)
-    elif fluid_str == "PG":
-        inst = PropyleneGlycol(x_conc)
-    elif fluid_str == "W":
-        inst = Water()
+@click.command(name="SecondaryCoolantPropsCommandLine")
+@click.option('-f', '--fluid', type=fluid_options, required=True, default='water', help="Which fluid to use?")
+@click.option('-x', '--concentration', type=x_range, required=False, default=0.0, help="Mixture concentration fraction. Default 0.0.")
+@click.option('-p', '--property', 'fluid_prop', type=prop_options, required=True, help="Which fluid property to evaluate.")
+@click.option('-t', '--temperature', type=float, required=True, help="Fluid temperature, in degrees Celsius.")
+@click.option('-q', "--quick", is_flag=True, show_default=True, default=False, help="Just report the value, good for scripts")
+def cli(fluid: str, concentration: float, fluid_prop: str, temperature: float, quick: bool):
+    if concentration == 0.0 and fluid != 'water':
+        print("Mixture requested, but concentration zero, assuming water and continuing", file=stderr)
+    elif concentration > 0.0 and fluid == 'water':
+        print("Pure water requested, but nonzero concentration entered, assuming water and continuing", file=stderr)
+    if fluid == 'water' or concentration == 0.0:
+        f = water.Water()
     else:
-        inst = None
-        raise NotImplementedError("Should never get here")
-
-    if prop_str == "V":
-        return inst.viscosity(temp)
-
-    if prop_str == "C":
-        return inst.specific_heat(temp)
-
-    if prop_str == "D":
-        return inst.density(temp)
-
-    if prop_str == "L":
-        return inst.conductivity(temp)
-
-
-def main():
-    """
-    Primary calling point for console usage.
-
-    Requires four positional arguments be passed:
-    PropStr - Property type string
-    Temp - Temperature in Celsius
-    XConc - Mixture concentration fraction, from 0 to 1
-    FluidStr - Fluid type string
-    """
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("PropStr", type=str, help=f"Property strings: {_valid_prop_strs}")
-    parser.add_argument("Temp", type=float, help="Fluid temperature, in Celsius: [<100]")
-    parser.add_argument("XConc", type=float, help="Concentration fraction: [0-1]")
-    parser.add_argument("FluidStr", type=str, help=f"Fluid strings: {_valid_fluid_strs}")
-    args = parser.parse_args()
-    print(SCProps(args.PropStr, args.Temp, args.XConc, args.FluidStr))
-
-
-if __name__ == "__main__":
-    sys.exit(main())
+        f = fluids[fluid](concentration)
+    value = getattr(f, fluid_prop)(temperature)
+    units = getattr(f, f"{fluid_prop}_units")()
+    if quick:
+        print(value)
+    else:
+        print(f"Fluid:    {fluid}\nProperty: {fluid_prop}\nValue:    {value}\nUnits:    [{units}]")
